refactor: split agent loop and abstract model client
