+++ src/lib/tools/read-page.ts
@@ -6,6 +6,7 @@ import type {
 } from "@/lib/tools/types";
 import { launchPlaywrightChromium } from "@/lib/tools/playwright-browser";
 import { assertSafeUrl } from "@/lib/tools/url-guard";
+import { formatReadPageAnswer } from "@/lib/agent/page-result-format";
 
 const NAVIGATION_TIMEOUT_MS = 15_000;
 const MAX_TEXT_LENGTH = 2_000;
@@ -208,4 +209,21 @@ export const readPageTool: AgentTool = {
     additionalProperties: false,
   },
   execute: executeReadPage,
+  onResult(goal, result) {
+    if (!result.ok) {
+      return {
+        type: "immediate",
+        message: result.content,
+      };
+    }
+
+    const formattedReadPageAnswer = formatReadPageAnswer(goal, result.content);
+
+    return formattedReadPageAnswer
+      ? {
+          type: "immediate",
+          message: formattedReadPageAnswer,
+        }
+      : null;
+  },
 };
